requirements.txt
mido
python-dotenv
openai
